requirements.txt
ultralytics>=8.1.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.0.0
sahi>=0.11.0
tqdm>=4.66.0
scikit-learn>=1.3.0
